fix: scope the highlight disposable cache to the connection
Fixes #5211
